fix(ci): peel email helpers to sibling modules; denylist rules without RegExp; title-level fold guard; registries
- module-size: the email-thread helpers, the save-time slug canonicalizer and
  the --model gate move to src/commands/extract-conversation-facts/{email,
  entity-slug-canonicalizer,model-flag}.ts (re-exported by the command, no
  import path changes). Ceilings set to the measured sizes with notes for the
  remaining growth (extract-conversation-facts +309, jobs +10, engines +2/+3,
  types +11, config +4).
- semgrep detect-non-literal-regexp: `facts.email_automated_senders` entries
  are plain strings (an entry with `@` matches the address exactly or as a
  suffix, any other entry as a substring) and are never compiled into a
  RegExp; built-in rules stay literal regexes.
- eval conversation-parser: a heading above the pattern's anchor level (the
  `# Email thread:` title over `##` anchors) is a document title, not a
  folded speaker, so the fixture reports no unrecognized_headings.
- cli-flag-registry.generated.ts and structural-suites.tsv regenerated
  (--background/--follow no-ops, --model, the new test suites).
